Rename SiteStatus namespace to SystemStatus
This matches the name of the Drupal module used by the project.
