+++ src/SystemStatus/Data.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DrupalSecurityJira\SiteStatus;
+namespace DrupalSecurityJira\SystemStatus;
 
 interface Data
 {
