+++ src/SystemStatus/UnencryptedData.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DrupalSecurityJira\SiteStatus;
+namespace DrupalSecurityJira\SystemStatus;
 
 class UnencryptedData implements Data
 {
